Refs #128. Split up the test_styles module
